requirements.txt
alabaster==1.0.0
asgiref==3.8.1
babel==2.17.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.5
django-allauth==65.3.1
django-unfold==0.46.0
docutils==0.21.2
et_xmlfile==2.0.0
idna==3.10
imagesize==1.4.1
Jinja2==3.1.5
jwt==1.3.1
MarkupSafe==3.0.2
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
pockets==0.9.1
pycparser==2.22
Pygments==2.19.1
PyJWT==1.7.1
python-http-client==3.3.7
reportlab==4.2.5
requests==2.32.3
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.17.0
snowballstemmer==2.2.0
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
urllib3==2.3.0
